streamlit_app/requirements.txt
boto3==1.34.113
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
python_steam_api==2.0
Requests==2.32.3
streamlit==1.35.0
xgboost==1.7.1
scikit-learn==1.5.0
